+++ src/EntityType.h
@@ -28,7 +28,7 @@ public:
 	EntityType();
 	EntityType(SDL_RWops *stream);
 	virtual ~EntityType() { }
-	virtual const char **getVars(int *size, Uint32 **data)=0;
+	virtual const char **getVars(size_t *size, Uint32 **data)=0;
 	virtual void init(void);	
 	virtual void load(SDL_RWops *stream);
 	virtual bool loadText(SDL_RWops *stream);
