+++ src/UnitType.h
@@ -40,8 +40,8 @@ public:
 	UnitType() { init(); }
 	UnitType(SDL_RWops *stream) { load(stream); }
 	virtual ~UnitType() {}
-	virtual const char **getVars(int *size, Uint32 **data)
-	{	
+	virtual const char **getVars(size_t *size, Uint32 **data)
+	{
 		static const char *vars[] =
 		{
 			"startImageStopWalk",
